+++ crates/swc/tests/fixture/issues-1xxx/1279/case1/output/index.ts
@@ -1,10 +1,9 @@
-import * as swcHelpers from "@swc/helpers";
 export class Foo {
     nested() {
         let Foo1 = class Foo {
         };
-        swcHelpers.defineProperty(Foo1, "foo", "foo");
-        swcHelpers.defineProperty(Foo1, "bar", Foo1.foo);
+        Foo1.foo = "foo";
+        Foo1.bar = Foo1.foo;
         return new Foo1();
     }
 }